Moved project files into own folder. Started extracting heavily nested code.
