+++ src/projects/Project.ts
@@ -1,5 +1,5 @@
 import { ChannelFlags, DefaultReactionEmoji, ForumChannel, GuildForumThreadMessageCreateOptions, MessageEditOptions, PermissionsBitField } from "discord.js";
-import CrossingGuardBot from "./CrossingGuardBot";
+import CrossingGuardBot from "../CrossingGuardBot";
 import ProjectAttachment from "./ProjectAttachment";
 import ProjectLink from "./ProjectLink";
 import ProjectStaff from "./ProjectStaff";
